+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,18 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Added clinical-lowlight-theme theming files
+
+5.41.0 - (February 13, 2020)
+------------------
+### Fixed
+* Fixed Option Group labels not read on desktop screen readers.
+
+5.40.0 - (January 28, 2020)
+------------------
+### Changed
+* Minor dependency version bump
 
 5.39.0 - (January 7, 2020)
 ------------------